map page reorganize
