+++ src/App.tsx
@@ -77,5 +77,4 @@ function App() {
     </AppProvider>
   );
 }
-
 export default App;
