+++ src/components/ItineraryPanel/index.css
@@ -3,7 +3,6 @@
   height: 100%;
   background-color: white;
   border-left: 1px solid var(--color-neutral-200);
-  box-shadow: -2px 0 10px rgba(0, 0, 0, 0.05);
   overflow-y: auto;
   transition: transform var(--transition-normal), width var(--transition-normal);
 }
