+++ src/components/NavbarColumn/index.css
@@ -1,0 +1,37 @@
+.navbar-column {
+  width: 200px;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  padding: 20px;
+  align-items: center;
+}
+
+.navbar-logo {
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  gap: var(--space-2);
+  padding-bottom: var(--space-4);
+}
+
+.navbar-logo h1 {
+  font-size: 1.5rem;
+  font-weight: 700;
+  color: var(--color-neutral-800);
+}
+
+.navbar-column button {
+  background: transparent;
+  border: none;
+  font-weight: 500;
+  padding: 8px 16px;
+  transition: color 0.2s ease;
+  width: 10rem;
+  align-items: center;
+}
+
+.navbar-column button:hover {
+  color: #1a4b6e;
+}
